pyproject.toml
[project]
name = "data-movies"
version = "0.1.0"
description = ""
authors = [
    {name = "Daria",email = "digreen17@mail.ru"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pandas (>=2.2.3,<3.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "cookiecutter-data-science (>=2.2.0,<3.0.0)",
    "kaggle (>=1.7.4.5,<2.0.0.0)"
]

[tool.poetry]

[tool.poetry.group.dev.dependencies]
flake8 = "^7.2.0"
black = "^25.1.0"
isort = "^6.0.1"
mypy = "^1.15.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
